api/requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
blinker==1.4
cachelib==0.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Deprecated==1.2.13
faiss-cpu==1.7.2
Fiona==1.8.21
Flask==2.0.2
flask-redis==0.4.0
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
geographiclib==1.52
geopandas==0.11.0
geopy==2.2.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
Js2Py==0.71
line-profiler==3.5.1
MarkupSafe==2.0.1
munch==2.5.0
numpy==1.23.0
packaging==21.3
pandas==1.4.3
py4j==0.10.9.5
pyjsparser==2.7.1
pyparsing==3.0.7
pyproj==3.3.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.3.3
requests==2.27.1
Rtree==1.0.0
scikit-learn==1.0.2
scipy==1.8.0
sentry-sdk==1.7.1
Shapely==1.8.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.31
threadpoolctl==3.1.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.0.2
wrapt==1.14.1
